feat: simple expr casting
