+++ exact-numbers/src/test/kotlin/xyz/lbres/exactnumbers/expressions/expression/simple/castingTests.kt
@@ -1,0 +1,22 @@
+package xyz.lbres.exactnumbers.expressions.expression.simple
+
+fun runToByteTests() {
+}
+
+fun runToCharTests() {
+}
+
+fun runToShortTests() {
+}
+
+fun runToIntTests() {
+}
+
+fun runToLongTests() {
+}
+
+fun runToFloatTests() {
+}
+
+fun runToDoubleTests() {
+}
